refactor: fix clippy warnings and unsafe patterns
- state_impl.rs: 7x caps.get(n).unwrap() -> safe and_then
- async_ops.rs: collapsible_if, filter->find, last->next_back, map_or->is_some_and, let_underscore_future
- sync_ops.rs: manual char comparison, manual strip_prefix
- metadata.rs: &Path::new(&x) -> Path::new(&x)
- serve_api.rs: io::Error::other, removed unnecessary let
- hardware.rs/tls.rs/web_search.rs/models.rs: collapsible_if
- main.rs: collapsible_match
- event/panel/models.rs: manual char comparison
- event/overlay/mod.rs: vec_init_then_push
- render/overlays.rs: unnecessary parens
- models.rs: from_str -> parse_backend to avoid std trait confusion

Clippy warnings reduced from 117 to 21 (all remaining are skipped by design).

+++ src/tui/render/hints.rs
@@ -257,7 +257,7 @@ pub fn render_hints_line(app: &App, panel_area: ratatui::layout::Rect) -> Line<'
     let mut total_width: usize = 0;
 
     for span in &hints {
-        let span_width = span.width() as usize;
+        let span_width = span.width();
         if total_width + span_width <= available_width {
             total_width += span_width;
             truncated.push(span.clone());
